+++ packages/mui/src/~components/MuiTabs.tsx
@@ -4,9 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Role } from "@ariakit/react/role";
+import useMediaQuery from "@mui/material/useMediaQuery";
 import {
 	type BaseProps,
 	forwardRef,
+	useEventHandlers,
 } from "@stratakit/foundations/secret-internals";
 
 import type Tabs from "@mui/material/Tabs";
@@ -30,6 +32,27 @@ const MuiTabs = forwardRef<"div", MuiTabsProps>((props, forwardedRef) => {
 });
 DEV: MuiTabs.displayName = "MuiTabs";
 
+const MuiTab = forwardRef<"button", BaseProps<"button">>(
+	(props, forwardedRef) => {
+		const motionOk = useMediaQuery("(prefers-reduced-motion: no-preference)");
+
+		return (
+			<Role.button
+				{...props}
+				onFocus={useEventHandlers(props.onFocus, (event) => {
+					event?.currentTarget?.scrollIntoView({
+						block: "nearest",
+						inline: "nearest",
+						behavior: motionOk ? "smooth" : "auto",
+					});
+				})}
+				ref={forwardedRef}
+			/>
+		);
+	},
+);
+DEV: MuiTab.displayName = "MuiTab";
+
 // ----------------------------------------------------------------------------
 
-export { MuiTabs };
+export { MuiTab, MuiTabs };